Modified endpoints path for question retrieval and answering
