+++ backend/endpoints/teaching/__init__.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter
-from . import random
+from . import random, test
 
 router = APIRouter()
 router.include_router(random.router, prefix="/random")
+router.include_router(test.router, prefix="/test")
